Fix Authorization Matrix property support on jobs in a folder (v2)
When a job is stored in a folder and the JJB definition file uses "name"
key to specify the full path, folder + the actual name of the job, the
authorization matrix property is missing <inheritanceStrategy> tag in
the output XML.

This fix is an extension of a fix I made here:
- https://github.com/romanek-adam/jenkins-job-builder/commit/074985c7ff9360bb58be80ffab686746267f814f

Instead of just checking for "folder" key we now also check for a folder
separator in the "name" key to determine if a given job is stored in a
folder. This way we support both cases and generate the missing
<inheritanceStrategy> tag.

The copyright reflects this and the previous contribution in this
module.

Change-Id: I84b22c09c8a107aab2b4eca20feffc9b61675a92